# Improve performance of some Complex methods where call Numeric#real? …